refactor: move pystrukts folder to root, delete src
Move the source code from /src/pystrukts to /pystrukts
