+++ .github/workflows/lint.yml
@@ -16,4 +16,4 @@ jobs:
         
     - name: Analysing the code with pylint
       run: |
-        pylint src/pystrukts
+        pylint pystrukts
