+++ pystrukts/__init__.py
@@ -1,0 +1,4 @@
+# pylint: skip-file
+
+from .max_heap import MaxHeap
+from .union_find import UnionFind
