+++ pystrukts/list/__init__.py
@@ -1,0 +1,4 @@
+# pylint: skip-file
+
+from .singly_linked_list import SinglyLinkedList, SinglyLinkedNode
+from .doubly_linked_list import DoublyLinkedList, DoublyLinkedNode
