+++ tests/test_singly_linked_list.py
@@ -1,5 +1,5 @@
 import pytest
-from src.pystrukts.list import SinglyLinkedList, SinglyLinkedNode
+from pystrukts.list import SinglyLinkedList, SinglyLinkedNode
 
 def test_node_initialization():
     node = SinglyLinkedNode(1)
